Websrv_lib, http_server: have special exception logging
the websrv_lib must conform to the BaseHTTPServer API, that does not
define an exception logging function, nor does it encourage the use
of individual pythonic loggers. So, have a function that is optionally
called, upon exception.

This improves much over the 'log_error()' call, because it allows us
to see all those bugs that would make a http method fail.