[FIX] report_sxw: fix printing odt/sxw reports after regression from fix for bug 670778
The original fix introduced a safety mechanism for avoiding transmission of invalid binary data via
XML-RPC, but caused an issue for binary function fields that do not carry data as base64.
This patch fixes this problem with the printing of odt/sxw reports (via odt2odt) which does
rely on non-base64 binary data for the sxw content.

By re-decoding the unicode-wrapped binary data using the pass-through charset latin1 we
are sure to restore the original byte values, hence the original file content.