[FIX] res_user: allow user to switch to any company she belongs to.
Due to global ir.rules applied on res.company, a user should only
be able to see the child companies of the company he is currently
working in. However this needs to be bypassed when the user is
changing to a different company through user preferences.
It is safe to perform this check using super-user to bypass all
ir.rules.