[FIX] db dump: report errors due to missing password + pass PGPASSWORD even on Unix
Even on Unix systems the PGPASSWORD env var
is necessary when Postgres is configured to
deny unix socket connections or to require
a password anyway. This avoids asking admins
to manually create a ~/.pgpass file in order
to use the dump/restore tools.
Also did some cleanup to ensure the PGPASSWORD
var is removed from the environment as soon
as the operation is done.